+++ packages/anthropic/src/anthropic-api-types.ts
@@ -21,12 +21,30 @@ export interface AnthropicUserMessage {
 
 export interface AnthropicAssistantMessage {
   role: 'assistant';
-  content: Array<AnthropicTextContent | AnthropicToolCallContent>;
+  content: Array<
+    | AnthropicTextContent
+    | AnthropicThinkingContent
+    | AnthropicRedactedThinkingContent
+    | AnthropicToolCallContent
+  >;
 }
 
 export interface AnthropicTextContent {
   type: 'text';
   text: string;
+  cache_control: AnthropicCacheControl | undefined;
+}
+
+export interface AnthropicThinkingContent {
+  type: 'thinking';
+  thinking: string;
+  signature: string;
+  cache_control: AnthropicCacheControl | undefined;
+}
+
+export interface AnthropicRedactedThinkingContent {
+  type: 'redacted_thinking';
+  data: string;
   cache_control: AnthropicCacheControl | undefined;
 }
 
@@ -74,18 +92,18 @@ export type AnthropicTool =
     }
   | {
       name: string;
-      type: 'computer_20241022';
+      type: 'computer_20250124' | 'computer_20241022';
       display_width_px: number;
       display_height_px: number;
       display_number: number;
     }
   | {
       name: string;
-      type: 'text_editor_20241022';
+      type: 'text_editor_20250124' | 'text_editor_20241022';
     }
   | {
       name: string;
-      type: 'bash_20241022';
+      type: 'bash_20250124' | 'bash_20241022';
     };
 
 export type AnthropicToolChoice =